Safari-611.1.10/LayoutTests/http/tests/workers/service/install-fails-expected.txt

PASS: Worker became redundant
PASS: registration.installing should be null
